refactor(weather): dedup getMockWeather across web and RN runtimes
Factor the platform-agnostic derivation logic (temp / humidity / wind / conditionIndex) into a shared mock-weather.ts module consumed by both runtimes. Each runtime keeps its own platform-specific condition representation (web semantic key + WeatherIcon, RN emoji glyph) and maps it from the shared conditionIndex at the call site, per issue #147's suggested fix. No behavior change. Rebased onto upstream/main to pick up the English-only source refactor (#148).

Closes #147
